Separate scf/nscf inputs and move defaults to protocol
Splitting up the inputs for the two PwBaseWorkChains is more consistent
with our approach e.g in the PwRelaxWorkChain and allows for full
customization of the inputs. It also avoids the situation where a user
who wants to specify an input for the scf step but not for the nscf step
needs to add it to the `base` input and then set the default to the
`nscf` input. All in all it's a clear way of defining the inputs for
these two steps of the work chain. Since we have now added the
`get_builder_from_protocol` method, it is no longer tedious to have to
specify the inputs twice.

We also move the default values to the protocol.yaml file of the work
chain. This means that a user that relies on the
get_builder_from_protocol method will have the correct inputs for the
scf and nscf steps of the work chain. However, we still have to add
validators to make sure that the inputs for each of these namespaces are
correct.