Add option to use `NUMBER_OF_QPOINTS` setting
To restart from a parent `ph.x` folder, the `epw.x` calcjob currently assumes
has a creator, and extracts the number of `qpoints` from the `output_parameters`
of this `PhCalculation`. However, we might want to restart from a remote folder
that does not have a creator in the database. Since we do not have an approach for
determining the number of q-points on the fly in this case, we add a setting that
allows the user to specify this.